refactor: extract text formatting utilities to src/utils/textFormatting.jsx
